backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
opencv-python>=4.9.0
numpy>=1.26.0
python-multipart>=0.0.9
pydantic>=2.7.0
websockets>=12.0
ultralytics>=8.0.0
PyTurboJPEG>=1.7.5
